Replaced Map of Subjects with a single subject
Map of subjects had an issue with a race condition.
Additionally already created subjects were never cleaned.
Replacing them with a single subject ensures that when
a subscriber finishes the subscrition is being disposed.
Keeping all the logic handling in a single monad makes sure
that all race conditions are handeled properly.