Add FETCHING_COMPLETED as a status for DataStreamNotification
The problem with no completed status is that the DataStreamNotification status
stream is completely separated from the value stream. As a consequence, there
is no guarantee that a completed query results in update in value stream. This
could happen e.g. in case a newly fetched value is identical with old value
already in the stream.

Solution is to handle status stream onCompleted and value stream onNext
separately.